+++ benchmarking_detectors/detector/watermark_detector.py
@@ -22,7 +22,7 @@ class WatermarkDetector(Detector):
         self.watermarking_scheme = watermarking_scheme
         self.detection_threshold = detection_threshold
         
-    def detect(self, texts: list[str], batch_size: int, detection_threshold: int) -> tuple[list[int], list[float], list[int]]:
+    def detect(self, texts: list[str], batch_size: int, detection_threshold: float) -> tuple[list[int], list[float], list[int]]:
         """
         Detect the if the texts given as input are watermarked (label 1) or not (label 0).
         """
